+++ src/Borm/Model/Construction/EntityBuilder.cs
@@ -1,4 +1,5 @@
 ﻿using Borm.Model.Validation;
+using Borm.Model.Validation.Expressions;
 using Borm.Reflection;
 
 namespace Borm.Model.Construction;
